PHP 8.0 compatibility: remove spacing within namespaced names
As of PHP 8.0, whitespace is no longer allowed within a namespaced name.

The Peast project even got a mention in the RFC in which this change was proposed as one of the very few projects which would be affected by this change.

This commit fixes the issue.

Note: as import `use` statement _always_ have to use fully qualified names, a leading backslash is unnecessary and strongly discouraged, though it won't cause a parse error (yet).

With that in mind, I've removed the leading backslashes in the files affected by the PHP 8 change as well, but haven't touched any of the other files containing import `use` statements with leading backslashes.

I'd encourage you to remove those as well though, but that is outside the scope of this PR.

Ref: https://wiki.php.net/rfc/namespaced_names_as_token